Weekend_Planner/weekend_planner_crew/requirements.txt
crewai
crewai-tools
browser_use
asyncio
python-dotenv
langchain_google_genai